+++ app/src/main/java/de/pawcode/cardstore/ui/components/EditCardForm.kt
@@ -283,9 +283,16 @@ fun EditCardForm(
               )
             },
             leadingIcon = {
+<<<<<<< HEAD
               AnimatedVisibility(visible = chipSelected) {
                 Icon(painterResource(R.drawable.check_solid), contentDescription = null)
               }
+=======
+              Icon(
+                painterResource(if (chipSelected) R.drawable.check_solid else R.drawable.add_solid),
+                contentDescription = null,
+              )
+>>>>>>> origin/staging
             },
           )
         }