+++ app/src/main/java/de/pawcode/cardstore/ui/screens/EditLabelScreen.kt
@@ -1,5 +1,6 @@
 package de.pawcode.cardstore.ui.screens
 
+import android.widget.Toast
 import androidx.activity.compose.BackHandler
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -27,6 +28,7 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.ImeAction
@@ -35,12 +37,11 @@ import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
-import androidx.navigation.NavController
 import de.pawcode.cardstore.R
 import de.pawcode.cardstore.data.database.entities.EXAMPLE_LABEL
 import de.pawcode.cardstore.data.database.entities.LabelEntity
 import de.pawcode.cardstore.data.database.entities.emptyLabel
-import de.pawcode.cardstore.data.services.SnackbarService
+import de.pawcode.cardstore.navigation.Navigator
 import de.pawcode.cardstore.ui.components.AppBar
 import de.pawcode.cardstore.ui.components.SaveFabComponent
 import de.pawcode.cardstore.ui.dialogs.UnsavedChangesDialog
@@ -48,10 +49,11 @@ import de.pawcode.cardstore.ui.viewmodels.CardViewModel
 
 @Composable
 fun EditLabelScreen(
-  navController: NavController,
+  navigator: Navigator,
   labelId: String? = null,
   viewModel: CardViewModel = viewModel(),
 ) {
+  val context = LocalContext.current
   rememberCoroutineScope()
 
   val initialLabel =
@@ -68,7 +70,7 @@ fun EditLabelScreen(
     EditLabelScreenComponent(
       isCreateLabel = labelId == null,
       initialLabel = initialLabel,
-      onBack = { navController.popBackStack() },
+      onBack = { navigator.goBack() },
       onSave = { label ->
         if (labelId != null) {
           viewModel.updateLabel(label)
@@ -76,9 +78,9 @@ fun EditLabelScreen(
           viewModel.insertLabel(label)
         }
 
-        navController.popBackStack()
+        navigator.goBack()
 
-        SnackbarService.showSnackbar(message = snackbarMessage)
+        Toast.makeText(context, snackbarMessage, Toast.LENGTH_SHORT).show()
       },
     )
   }